fix: expose drepIdToCredential
