+++ .changeset/quiet-knives-end.md
@@ -1,0 +1,6 @@
+---
+"@lucid-evolution/lucid": patch
+"@lucid-evolution/utils": patch
+---
+
+expose drepIDToCredential
