+++ packages/utils/package.json
@@ -34,6 +34,7 @@
   "dependencies": {
     "@anastasia-labs/cardano-multiplatform-lib-browser": "6.0.2-3",
     "@anastasia-labs/cardano-multiplatform-lib-nodejs": "6.0.2-3",
+    "@cardano-sdk/core": "^0.45.1",
     "@effect/schema": "^0.68.16",
     "@harmoniclabs/plutus-data": "^1.2.4",
     "@harmoniclabs/uplc": "^1.2.4",
